Printing Paths Error Fix
Functions that would print a path of lowest total weight or shortest length would cause an error, whenerver the path doesn't exist, in which the answer array gets filled with incorrect values and a large number of elements are printed. In order to avoid that, those functions no longer analyze more than a limited number of positions, considering the size of the graph. Now, all functions that print paths return an integer 0 or 1, describing if there really is a path between the vertices specified. As part of this commit, those functions have also had malloc size fixed (...*(size+1) istead of ...* size+1), and answer arrays are now being freed.